// src/types/pricing.ts
export type Currency = 'RUB' | 'USD' | 'EUR' | 'VND';
// ...
export interface PricingOptions {
  isClubMember: boolean;
  targetCurrency: Currency;
}